demo/chat/server/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
openai==1.1.0
